[maui] improve generated code
