+++ src/skiasharp/LiveChartsCore.SkiaSharpView.Maui/CartesianChart.cs
@@ -20,36 +20,17 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-// ==============================================================================
-// 
-// this file contains the MAUI specific code for the CartesianChart class,
-// the rest of the code can be found in the _Shared project.
-// 
-// ==============================================================================
-
-using LiveChartsCore.Drawing;
 using LiveChartsCore.Kernel.Sketches;
-using LiveChartsCore.Measure;
-using Microsoft.Maui.Controls.Xaml;
 
 namespace LiveChartsCore.SkiaSharpView.Maui;
 
-/// <inheritdoc cref="ICartesianChartView"/>
-[XamlCompilation(XamlCompilationOptions.Compile)]
-public partial class CartesianChart : ChartControl, ICartesianChartView
-{
-    internal override void OnScrolled(object? sender, Native.Events.ScrollEventArgs args)
-    {
-        var c = (CartesianChartEngine)CoreChart;
-        c.Zoom(ZoomMode, args.Location, args.ScrollDelta > 0 ? ZoomDirection.ZoomIn : ZoomDirection.ZoomOut);
-    }
+// ==============================================================================
+// 
+// use the LiveChartsGeneratedCode.CartesianChart class to add maui specific
+// code, this class is just to expose the CartesianChart class in this namespace.
+// 
+// ==============================================================================
 
-    internal override void OnPinched(object? sender, Native.Events.PinchEventArgs args)
-    {
-        var c = (CartesianChartEngine)CoreChart;
-        var p = args.PinchStart;
-        var s = c.ControlSize;
-        var pivot = new LvcPoint((float)(p.X * s.Width), (float)(p.Y * s.Height));
-        c.Zoom(ZoomMode, pivot, ZoomDirection.DefinedByScaleFactor, args.Scale);
-    }
-}
+/// <inheritdoc cref="ICartesianChartView"/>
+public partial class CartesianChart : LiveChartsGeneratedCode.CartesianChart
+{ }
