+++ src/skiasharp/LiveChartsCore.SkiaSharpView.Maui/Handlers/EmptyContentView.cs
@@ -1,4 +1,4 @@
-// The MIT License(MIT)
+﻿// The MIT License(MIT)
 //
 // Copyright(c) 2021 Alberto Rodriguez Orozco & LiveCharts Contributors
 //
@@ -20,12 +20,44 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using Microsoft.Maui.Handlers;
+using Microsoft.Maui;
+using Microsoft.Maui.Controls;
+using Microsoft.Maui.Graphics;
 
-namespace LiveChartsCore.SkiaSharpView.Maui;
+namespace LiveChartsCore.SkiaSharpView.Maui.Handlers;
 
 /// <summary>
-/// An empty <see cref="ContentViewHandler"/> that does nothing.
+/// Just a view that does nothing, but helps to make the XAML work.
 /// </summary>
-public class EmptyViewHandler : ContentViewHandler
-{ }
+public class EmptyContentView : View, IContentView
+{
+    /// <summary>
+    /// Gets the content.
+    /// </summary>
+    public object? Content => null;
+
+    /// <summary>
+    /// Gets the presented content.
+    /// </summary>
+    public IView? PresentedContent => null;
+
+    /// <summary>
+    /// Gets the padding.
+    /// </summary>
+    Thickness IPadding.Padding => new();
+
+    /// <summary>
+    /// Gets the size.
+    /// </summary>
+    /// <param name="bounds"></param>
+    /// <returns></returns>
+    public Size CrossPlatformArrange(Rect bounds) => new();
+
+    /// <summary>
+    /// Measures the specified width constraint.
+    /// </summary>
+    /// <param name="widthConstraint"></param>
+    /// <param name="heightConstraint"></param>
+    /// <returns></returns>
+    public Size CrossPlatformMeasure(double widthConstraint, double heightConstraint) => new();
+}
