+++ src/skiasharp/_Shared.Xaml/CartesianAxes.cs
@@ -34,12 +34,13 @@ using System.Collections.Generic;
 using LiveChartsCore.Generators;
 using LiveChartsCore.Kernel.Sketches;
 
+
 #if AVALONIA_LVC
 using Avalonia;
 using BaseControl = Avalonia.Controls.Control;
 namespace LiveChartsCore.SkiaSharpView.Avalonia;
 #elif MAUI_LVC
-using BaseControl = LiveChartsCore.SkiaSharpView.Maui.EmptyContentView;
+using BaseControl = LiveChartsCore.SkiaSharpView.Maui.Handlers.EmptyContentView;
 namespace LiveChartsCore.SkiaSharpView.Maui;
 #elif WINUI_LVC
 using BaseControl = Microsoft.UI.Xaml.FrameworkElement;
