+++ src/skiasharp/_Shared.Xaml/PolarAxes.cs
@@ -23,11 +23,12 @@
 using LiveChartsCore.Generators;
 using LiveChartsCore.Kernel.Sketches;
 
+
 #if AVALONIA_LVC
 using BaseControl = Avalonia.Controls.Control;
 namespace LiveChartsCore.SkiaSharpView.Avalonia;
 #elif MAUI_LVC
-using BaseControl = LiveChartsCore.SkiaSharpView.Maui.EmptyContentView;
+using BaseControl = LiveChartsCore.SkiaSharpView.Maui.Handlers.EmptyContentView;
 namespace LiveChartsCore.SkiaSharpView.Maui;
 #elif WINUI_LVC
 using BaseControl = Microsoft.UI.Xaml.FrameworkElement;
